virtio: event suppression support for packed ring
Difference between 'avail_wrap_counter' and 'last_avail_wrap_counter':
For Tx(guest transmitting), they are the same after each pop of a desc.

For Rx(guest receiving), they are also the same when there are enough
descriptors to carry the payload for a packet(usually 16 descs are needed
for a 64k packet in typical tcp connection with tso enabled), however,
when the ring is running out of descriptors while there are still a few
free ones, e.g. 6 descriptors are available that is not enough to
carry an entire packet which needs 16 descriptors, then 'avail_wrap_counter'
should be set as the first pending one not handled by guest driver in next
order to get a notification, and 'last_avail_wrap_counter' should stay
unchanged at the head of available descriptors.

Signed-off-by: Wei Xu <wexu@redhat.com>