virtio-net: update head descriptor in a chain lastly
This is a helper for packed ring.

To support packed ring, the head descriptor in a chain should be updated
lastly since no 'avail_idx' like in packed ring to explicitly tell the
other side all payload is ready after having done the chain, so the head
 is always visible immediately.

This patch fills the head one after done all the others one.

Signed-off-by: Wei Xu <wexu@redhat.com>